refactor: Move lint_type.dart to models folder
