+++ tools/check_lint_rules_identity/lib/src/models/lint_type.dart
