+++ tools/check_lint_rules_identity/lib/src/version_data_source.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:async/async.dart';
 import 'package:check_lint_rules_identity/src/lint_rules_dir.dart';
-import 'package:check_lint_rules_identity/src/lint_type.dart';
+import 'package:check_lint_rules_identity/src/models/lint_type.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
